% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_household.R
\name{calculate_household}
\alias{calculate_household}
\title{A function to calculate the house-hold not home air pollution}
\usage{
calculate_household(act_dat, pop_dat, ambient, outvar)
}
\arguments{
\item{act_dat}{\itemize{
\item Dataset containing the act profiles
with associated ambient air pollution
}}

\item{pop_dat}{\itemize{
\item Population dataset with attributes
needed for the mass balance equations
}}

\item{ambient}{\itemize{
\item column containing the ambient air
pollution estimates
}}

\item{outvar}{\itemize{
\item Column to output the household concent-
rations to
}}
}
\value{
Dataset containing the sampled PM2.5 estimate
for household.
}
\description{
A function to calculate the house-hold not home air pollution
}
