% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indoor.R
\name{calculate_indoor}
\alias{calculate_indoor}
\title{A function to calculate the indoor not home air pollution}
\usage{
calculate_indoor(dat, ambient, outvar)
}
\arguments{
\item{dat}{\itemize{
\item Dataset containing sampled population with
indoor not home activities
}}
}
\value{
Dataset containing the sampled PM2.5 estimate
for indoor not home
}
\description{
A function to calculate the indoor not home air pollution
}
