% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_outdoor.R
\name{calculate_outdoor}
\alias{calculate_outdoor}
\title{A function to calculate the outdoor air pollution}
\usage{
calculate_outdoor(dat)
}
\arguments{
\item{dat}{\itemize{
\item Dataset containing sampled population with
outdoor activities
}}
}
\value{
Dataset containing the sampled PM2.5 estimate
for outdoors
}
\description{
A function to calculate the outdoor air pollution
}
