% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transport.R
\name{calculate_transport}
\alias{calculate_transport}
\title{A function to calculate the transportation air pollution}
\usage{
calculate_transport(dat, ambient, outvar)
}
\arguments{
\item{dat}{\itemize{
\item Dataset containing sampled population with
transportation activities
}}
}
\value{
Dataset containing the sampled PM2.5 estimate
for indoor not home
}
\description{
A function to calculate the transportation air pollution
}
