% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_averages.R
\name{daily_averages}
\alias{daily_averages}
\title{Estimate daily average personal exposure for a specific MSOA.}
\usage{
daily_averages(activities_complete)
}
\arguments{
\item{activities_complete}{\itemize{
\item sampled activities from earlier processing steps
}}
}
\value{
dataframe of personal daily average exposure estimates
}
\description{
Estimate daily average personal exposure for a specific MSOA.
}
