% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-helpers.R
\name{day_number}
\alias{day_number}
\title{Day of week in terms of a number}
\usage{
day_number(date)
}
\arguments{
\item{date}{\itemize{
\item a Date object
}}
}
\value{
integer between 1 and 7 inclusive
}
\description{
Day of week in terms of a number
}
