% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-helpers.R
\name{day_type}
\alias{day_type}
\title{Type of day in terms of a number}
\usage{
day_type(day)
}
\arguments{
\item{day}{\itemize{
\item integer between 1 and 7
}}
}
\value{
integer representing either week day or weekend
}
\description{
Type of day in terms of a number
}
