% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-helpers.R
\name{day_type_label}
\alias{day_type_label}
\title{Type of day, either Weekend or Weekday}
\usage{
day_type_label(day)
}
\arguments{
\item{day}{\itemize{
\item integer between 1 and 7
}}
}
\value{
string either "Weekend" or "Weekday"
}
\description{
Type of day, either Weekend or Weekday
}
