% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_exposures.R
\name{estimate_exposures}
\alias{estimate_exposures}
\title{Estimating exposure}
\usage{
estimate_exposures(activities_complete, pop_dat, pm25_ctm)
}
\arguments{
\item{activities_complete}{\itemize{
\item output data from activity sampler
}}

\item{pop_dat}{\itemize{
\item population data frame
}}

\item{pm25_ctm}{\itemize{
\item aggregated dataframe made up of EMEP and CAMS data
}}
}
\value{
data frame of completed activities
}
\description{
Estimating exposure
}
