% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_sampler_fill_in_missings.R
\name{fill_in_missings}
\alias{fill_in_missings}
\title{Filling in missings in the Time Use Survey dataset
using most popular activities}
\usage{
fill_in_missings(tus_dat)
}
\arguments{
\item{tus_dat}{\itemize{
\item Time use survey data frame that needs processing
}}
}
\value{
Time use survey data frame
}
\description{
Filling in missings in the Time Use Survey dataset
using most popular activities
}
