% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pollutants.R
\name{merge_pollutants}
\alias{merge_pollutants}
\title{Merge PM2.5 pollutant data from CAMS and EMEP models.}
\usage{
merge_pollutants(pm25_cams, pm25_emep, start_date, end_date)
}
\arguments{
\item{pm25_cams}{\itemize{
\item Dataset containing PM2.5 levels from CAMS
}}

\item{pm25_emep}{\itemize{
\item Dataset containing PM2.5 levels from EMEP
}}

\item{start_date}{\itemize{
\item Start date
}}

\item{end_date}{\itemize{
\item End date
}}
}
\value{
Dataset of merged PM2.5 pollutant levels
}
\description{
Merge PM2.5 pollutant data from CAMS and EMEP models.
}
