% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_group.R
\name{micro_group}
\alias{micro_group}
\title{Categorise micro-environment from location ID}
\usage{
micro_group(location)
}
\arguments{
\item{location}{\itemize{
\item ID indicating location
}}
}
\value{
micro-environment text representation
}
\description{
Categorise micro-environment from location ID
}
