% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_population.R
\name{sample_population}
\alias{sample_population}
\title{A function to sample activities sequences from the time use survey}
\usage{
sample_population(
  pop_dat,
  tus_dat,
  nsample,
  weights = NULL,
  pop_strata,
  tus_strata,
  start_date,
  end_date,
  keep
)
}
\arguments{
\item{pop_dat}{\itemize{
\item Population data to sample from
}}

\item{tus_dat}{\itemize{
\item Time use survey data to sample from
}}

\item{nsample}{\itemize{
\item Number of population samples
}}

\item{weights}{\itemize{
\item Sampling weights for the time use diaries
}}

\item{pop_strata}{\itemize{
\item Stratification variables for sampling
population data
}}

\item{tus_strata}{\itemize{
\item Stratification variables for sampling
time use data
}}

\item{start_date}{\itemize{
\item Start date for personal exposures
}}

\item{end_date}{\itemize{
\item End date for personal exposures
}}

\item{keep}{\itemize{
\item Variables to keep from the TUS
}}
}
\value{
Dataframe with sampled activity sequences for
each individual between start_date end_date
}
\description{
A function to sample activities sequences from the time use survey
}
