% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_population.R
\name{sample_sequences}
\alias{sample_sequences}
\title{Sample activity sequences}
\usage{
sample_sequences(activities, tus_act_id)
}
\arguments{
\item{activities}{\itemize{
\item data frame with strata, pop_id and act_id columns
}}

\item{tus_act_id}{\itemize{
\item data frame with strata and weights columns
}}
}
\value{
modified activities data frame
}
\description{
Use time use survey activities to populate table
of activities
}
\details{
Note: in this context strata is an integer index representation
of unique combinations calculated previously
}
