% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-helpers.R
\name{season}
\alias{season}
\title{Season}
\usage{
season(month)
}
\arguments{
\item{month}{\itemize{
\item integer between 1 and 12 inclusive
}}
}
\value{
integer representing season
}
\description{
Season
}
