% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-helpers.R
\name{season_label}
\alias{season_label}
\title{Season label}
\usage{
season_label(month)
}
\arguments{
\item{month}{\itemize{
\item integer between 1 and 12 inclusive
}}
}
\value{
string of "Winter", "Spring", "Summer" or "Autumn"
}
\description{
Season label
}
