% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_sampler.R
\name{activity_sampler}
\alias{activity_sampler}
\title{Sampling activity sequences (Method 1) - Complete sequences only}
\usage{
activity_sampler(pop_dat, tus_dat, k)
}
\arguments{
\item{pop_dat}{\itemize{
\item population data frame
}}

\item{tus_dat}{\itemize{
\item time-use survey data frame
}}

\item{k}{\itemize{
\item key to index chunk
}}
}
\value{
data frame of sampled activities
}
\description{
Sampling activity sequences (Method 1) - Complete sequences only
}
