% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-helpers.R
\name{date_filter}
\alias{date_filter}
\title{Date filter}
\usage{
date_filter(frame, start_date, end_date)
}
\arguments{
\item{frame}{\itemize{
\item data frame to filter
}}

\item{start_date}{\itemize{
\item start date of date filter interval (inclusive)
}}

\item{end_date}{\itemize{
\item end date of date filter interval (inclusive)
}}
}
\value{
only dates within filter window
}
\description{
Date filter
}
